+++ src/GoogleDriveBackupManager.php
@@ -17,7 +17,7 @@ class GoogleDriveBackupManager
      */
     public function version(): string
     {
-        return '2.0.0';
+        return '2.2.1';
     }
 
     /**